# contracts/api/orders.openapi.yaml
openapi: 3.0.3
info:
  title: Clothing Platform - Orders API
  description: API for managing orders, cart, payments, and shipping
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@clothingplatform.com

servers:
  - url: http://127.0.0.1:8000/api/orders
    description: Development server
  - url: https://api.clothingplatform.com/orders
    description: Production server

paths:
  /cart/:
    get:
      summary: Get cart contents
      operationId: getCart
      tags: [Cart]
      responses:
        '200':
          description: Cart contents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'

  /cart/add/:
    post:
      summary: Add item to cart
      operationId: addToCart
      tags: [Cart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToCartRequest'
      responses:
        '201':
          description: Item added to cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartItem'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /cart/remove/{item_id}/:
    delete:
      summary: Remove item from cart
      operationId: removeFromCart
      tags: [Cart]
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Item removed from cart
        '404':
          description: Cart item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /cart/item/{item_id}/:
    put:
      summary: Update cart item
      operationId: updateCartItem
      tags: [Cart]
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemRequest'
      responses:
        '200':
          description: Cart item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartItem'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Cart item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /checkout/:
    post:
      summary: Create order from cart
      operationId: checkout
      tags: [Checkout]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /orders/:
    get:
      summary: List user orders
      operationId: listOrders
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, processing, shipped, delivered, cancelled, refunded]
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrderList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create new order
      operationId: createOrder
      tags: [Orders]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /orders/{id}/:
    get:
      summary: Get order details
      operationId: getOrder
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /track/:
    get:
      summary: Track order by tracking number
      operationId: trackOrder
      tags: [Orders]
      parameters:
        - name: tracking_number
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order tracking information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderTracking'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /stats/:
    get:
      summary: Get order statistics
      operationId: getOrderStats
      tags: [Orders]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Order statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStats'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Cart:
      type: object
      properties:
        id:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        total_items:
          type: integer
        subtotal:
          type: number
          format: decimal
        tax_amount:
          type: number
          format: decimal
        shipping_amount:
          type: number
          format: decimal
        total_amount:
          type: number
          format: decimal
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - items
        - total_items
        - subtotal
        - tax_amount
        - shipping_amount
        - total_amount
        - created_at
        - updated_at

    CartItem:
      type: object
      properties:
        id:
          type: integer
        product:
          $ref: '#/components/schemas/CartProduct'
        quantity:
          type: integer
          minimum: 1
        size:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        unit_price:
          type: number
          format: decimal
        total_price:
          type: number
          format: decimal
        added_at:
          type: string
          format: date-time
      required:
        - id
        - product
        - quantity
        - unit_price
        - total_price
        - added_at

    CartProduct:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        price:
          type: number
          format: decimal
        sale_price:
          type: number
          format: decimal
          nullable: true
        image:
          type: string
          format: uri
          nullable: true
        stock_quantity:
          type: integer
        is_active:
          type: boolean
      required:
        - id
        - name
        - slug
        - price
        - stock_quantity
        - is_active

    AddToCartRequest:
      type: object
      properties:
        product_id:
          type: integer
        quantity:
          type: integer
          minimum: 1
          default: 1
        size:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
      required:
        - product_id

    UpdateCartItemRequest:
      type: object
      properties:
        quantity:
          type: integer
          minimum: 1
        size:
          type: string
          nullable: true
        color:
          type: string
          nullable: true

    CheckoutRequest:
      type: object
      properties:
        shipping_address_id:
          type: integer
        billing_address_id:
          type: integer
          nullable: true
        payment_method:
          type: string
          enum: [credit_card, debit_card, paypal, stripe, bank_transfer]
        payment_details:
          type: object
          additionalProperties: true
        coupon_code:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - shipping_address_id
        - payment_method

    CreateOrderRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'
        shipping_address_id:
          type: integer
        billing_address_id:
          type: integer
          nullable: true
        payment_method:
          type: string
          enum: [credit_card, debit_card, paypal, stripe, bank_transfer]
        payment_details:
          type: object
          additionalProperties: true
        coupon_code:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - items
        - shipping_address_id
        - payment_method

    OrderItemRequest:
      type: object
      properties:
        product_id:
          type: integer
        quantity:
          type: integer
          minimum: 1
        size:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
      required:
        - product_id
        - quantity

    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_number:
          type: string
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, refunded]
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        shipping_address:
          $ref: '#/components/schemas/OrderAddress'
        billing_address:
          $ref: '#/components/schemas/OrderAddress'
        payment_method:
          type: string
        payment_status:
          type: string
          enum: [pending, processing, completed, failed, refunded]
        subtotal:
          type: number
          format: decimal
        tax_amount:
          type: number
          format: decimal
        shipping_amount:
          type: number
          format: decimal
        discount_amount:
          type: number
          format: decimal
        total_amount:
          type: number
          format: decimal
        coupon_code:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        tracking_number:
          type: string
          nullable: true
        estimated_delivery:
          type: string
          format: date
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - order_number
        - status
        - items
        - shipping_address
        - billing_address
        - payment_method
        - payment_status
        - subtotal
        - tax_amount
        - shipping_amount
        - discount_amount
        - total_amount
        - created_at
        - updated_at

    OrderItem:
      type: object
      properties:
        id:
          type: integer
        product:
          $ref: '#/components/schemas/CartProduct'
        quantity:
          type: integer
        size:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        unit_price:
          type: number
          format: decimal
        total_price:
          type: number
          format: decimal
      required:
        - id
        - product
        - quantity
        - unit_price
        - total_price

    OrderAddress:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        company:
          type: string
          nullable: true
        address_line_1:
          type: string
        address_line_2:
          type: string
          nullable: true
        city:
          type: string
        state:
          type: string
        postal_code:
          type: string
        country:
          type: string
        phone:
          type: string
          nullable: true
      required:
        - first_name
        - last_name
        - address_line_1
        - city
        - state
        - postal_code
        - country

    OrderTracking:
      type: object
      properties:
        order_number:
          type: string
        tracking_number:
          type: string
        status:
          type: string
        estimated_delivery:
          type: string
          format: date
          nullable: true
        tracking_events:
          type: array
          items:
            $ref: '#/components/schemas/TrackingEvent'
      required:
        - order_number
        - tracking_number
        - status
        - tracking_events

    TrackingEvent:
      type: object
      properties:
        status:
          type: string
        description:
          type: string
        location:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
      required:
        - status
        - description
        - timestamp

    OrderStats:
      type: object
      properties:
        total_orders:
          type: integer
        total_spent:
          type: number
          format: decimal
        orders_by_status:
          type: object
          additionalProperties:
            type: integer
        recent_orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
      required:
        - total_orders
        - total_spent
        - orders_by_status
        - recent_orders

    PaginatedOrderList:
      type: object
      properties:
        count:
          type: integer
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/Order'
      required:
        - count
        - results

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - error
        - message